Fix unqualified exec is not allowed in function" SyntaxError
For some reason, I am seing this issue locally, while tests we passing on Travis.

To mitigate, moved new `namedtuple_gt_255_fields` out of the codeblock that is visible to Python 2.7 compiler.
